Extract table properties to a hook
